Fix Organizer Navigation Issues
